refactor to create checkout footer with total and checkout button
