+++ frontend/src/components/restaurant/index.tsx
@@ -3,3 +3,4 @@ export * from "./checkboxElement.tsx"
 export * from "./foodCard.tsx"
 export * from "./orders.tsx"
 export * from "./restaurantHeader.tsx"
+export * from "./checkoutFooter.tsx"
